refactor(sessions,cron): add SessionRouter interface — break cycle 1
shared/types.ts に SessionRouter・RouteOptions を追加。
cron/runner.ts・cron/scheduler.ts が sessions/manager.ts ではなく
shared/types.ts の SessionRouter を参照するよう変更。
sessions/manager.ts は RouteOptions を shared/types.ts から re-export。

madge --circular: 0 件達成。645 tests PASS。

TASK-020(ES-011) TASK-021(ES-011)
